refactor: CustomizableBrowserToolbar.swift
